Best solution I could come up with.. Probably not ideal but works
